Summary of generalized linear mixed models as HTML table

Summarizes (multiple) fitted generalized linear mixed models (odds ratios, ci, p-values...) as HTML table, or saves them as file. The fitted models may have different predictors, e.g. when comparing different stepwise fitted models.

sjt.glmer(..., pred.labels = NULL, depvar.labels = NULL,
  remove.estimates = NULL, group.pred = FALSE, exp.coef = TRUE,
  p.numeric = TRUE, emph.p = FALSE, p.zero = FALSE,
  separate.ci.col = TRUE, newline.ci = TRUE, show.ci = TRUE,
  show.se = FALSE, show.header = FALSE, show.col.header = TRUE,
  show.r2 = FALSE, show.icc = TRUE, show.re.var = TRUE,
  show.loglik = FALSE, show.aic = FALSE, show.aicc = FALSE,
  show.dev = TRUE, show.hoslem = FALSE, show.family = FALSE,
  string.pred = "Predictors", string.dv = "Dependent Variables",
  string.interc = "(Intercept)", string.obs = "Observations",
  string.est = NULL, string.ci = "CI", string.se = "std. Error",
  string.p = "p", ci.hyphen = "&nbsp;&ndash;&nbsp;", digits.est = 2,
  digits.p = 3, digits.ci = 2, digits.se = 2, digits.summary = 3,
  cell.spacing = 0.2, cell.gpr.indent = 0.6, sep.column = TRUE,
  CSS = NULL, encoding = NULL, file = NULL, use.viewer = TRUE,
  no.output = FALSE, remove.spaces = TRUE)

Arguments

...
One or more fitted generalized linear (mixed) models.

pred.labels
Character vector with labels of predictor variables. If not NULL, pred.labels will be used in the first table column with the predictors' names. If NULL, variable labels are set based on label attributes (see get_label). If pred.labels = "", column names (vector names) are used as predictor labels. See 'Examples'.

depvar.labels
Character vector with labels of dependent variables of all fitted models. See 'Examples'.

remove.estimates
Numeric vector with indices (order equals to row index of coef(fit)) or character vector with coefficient names that indicate which estimates should be removed from the table output. The first estimate is the intercept, followed by the model predictors. The intercept cannot be removed from the table output! remove.estimates = c(2:4) would remove the 2nd to the 4th estimate (1st to 3rd predictor after intercept) from the output. remove.estimates = "est_name" would remove the estimate est_name. Default is NULL, i.e. all estimates are printed.

group.pred
Logical, if TRUE (default), automatically groups table rows with factor levels of same factor, i.e. predictors of type factor will be grouped, if the factor has more than two levels. Grouping means that a separate headline row is inserted to the table just before the predictor values.

exp.coef
Logical, if TRUE (default), regression coefficients and confidence intervals are exponentiated. Use FALSE for non-exponentiated coefficients (log-odds) as provided by the summary function.

p.numeric
Logical, if TRUE, the p-values are printed as numbers. If FALSE (default), asterisks are used.

emph.p
Logical, if TRUE, significant p-values are shown bold faced.

p.zero
logical, if TRUE, p-values have a leading 0 before the period (e.g. 0.002), else p-values start with a period and without a zero (e.g. .002).

separate.ci.col
Logical, if TRUE, the CI values are shown in a separate table column. Default is FALSE.

newline.ci
Logical, if TRUE and separate.ci.col = FALSE, inserts a line break between estimate and CI values. If FALSE, CI values are printed in the same line as estimate values.

show.ci
Logical, if TRUE (default), the confidence intervall is also printed to the table. Use FALSE to omit the CI in the table.

show.se
Logical, if TRUE, the standard errors are also printed. Default is FALSE.

show.header
Logical, if TRUE, the header strings string.pred and string.dv are shown. By default, they're hidden.

show.col.header
Logical, if TRUE (default), the table data columns have a headline with abbreviations for estimates, std. beta-values, confidence interval and p-values.

show.r2
Logical, if TRUE (default), the pseudo R2 values for each model are printed in the model summary. R2cs is the Cox-Snell-pseudo R-squared value, R2n is Nagelkerke's pseudo R-squared value and D is Tjur's Coefficient of Discrimination (see cod).

show.icc
Logical, if TRUE, the intra-class-correlation for each model is printed in the model summary. Only applies to mixed models.

show.re.var
Logical, if TRUE, the variance parameters for the random effects for each model are printed in the model summary. Only applies to mixed models. For details output, see 'Note' in icc.

show.loglik
Logical, if TRUE, the Log-Likelihood for each model is printed in the model summary. Default is FALSE.

show.aic
Logical, if TRUE, the AIC value for each model is printed in the model summary. Default is FALSE.

show.aicc
Logical, if TRUE, the second-order AIC value for each model is printed in the model summary. Default is FALSE.

show.dev
Logical, if TRUE, the deviance for each model is printed in the model summary.

show.hoslem
Logical, if TRUE, a Hosmer-Lemeshow-Goodness-of-fit-test is performed. A well-fitting model shows no significant difference between the model and the observed data, i.e. the reported p-values should be greater than 0.05.

show.family
Logical, if TRUE, the family object and link function for each fitted model are printed. Can be used in case you want to compare models with different link functions and same predictors and response, to decide which model fits best. See family for more details. It is recommended to inspect the model AIC (see show.aic) to get a decision help for which model to choose.

string.pred
Character vector,used as headline for the predictor column. Default is "Predictors".

string.dv
Character vector, used as headline for the dependent variable columns. Default is "Dependent Variables".

string.interc
Character vector, used as headline for the Intercept row. Default is "Intercept".

string.obs
character vector, used in the summary row for the count of observation (cases). Default is "Observations".

string.est
Character vector, used for the column heading of estimates.

string.ci
Character vector, used for the column heading of confidence interval values. Default is "CI".

string.se
Character vector, used for the column heading of standard error values. Default is "std. Error".

string.p
Character vector, used for the column heading of p values. Default is "p".

ci.hyphen
Character vector, indicating the hyphen for confidence interval range. May be an HTML entity. See 'Examples'.

digits.est
Amount of decimals for table values.

digits.p
Amount of decimals for p-values.

digits.ci
Amount of decimals for confidence intervals.

digits.se
Amount of decimals for standard error.

digits.summary
Amount of decimals for values in model summary.

cell.spacing
Numeric, inner padding of table cells. By default, this value is 0.2 (unit is cm), which is suitable for viewing the table. Decrease this value (0.05 to 0.1) if you want to import the table into Office documents. This is a convenient argument for the CSS argument for changing cell spacing, which would be: CSS = list(css.thead = "padding:0.2cm;", css.tdata = "padding:0.2cm;").

cell.gpr.indent
Indent for table rows with grouped factor predictors. Only applies if group.pred = TRUE.

sep.column
Logical, if TRUE, an empty table column is added after each model column, to add margins between model columns. By default, this column will be added to the output; however, when copying tables to office applications, it might be helpful not to add this separator column when modifying the table layout.

CSS
A list with user-defined style-sheet-definitions, according to the official CSS syntax. For more details, see this package-vignette, or 'Details' in sjt.frq.

encoding
String, indicating the charset encoding used for variable and value labels. Default is NULL, so encoding will be auto-detected depending on your platform (e.g., "UTF-8" for Unix and "Windows-1252" for Windows OS). Change encoding if specific chars are not properly displayed (e.g. German umlauts).

file
Destination file, if the output should be saved as file. If NULL (default), the output will be saved as temporary file and openend either in the IDE's viewer pane or the default web browser.

use.viewer
Logical, if TRUE, the HTML table is shown in the IDE's viewer pane. If FALSE or no viewer available, the HTML table is opened in a web browser.

no.output
Logical, if TRUE, the html-output is neither opened in a browser nor shown in the viewer pane and not even saved to file. This option is useful when the html output should be used in knitr documents. The html output can be accessed via the return value.

remove.spaces
Logical, if TRUE, leading spaces are removed from all lines in the final string that contains the html-data. Use this, if you want to remove parantheses for html-tags. The html-source may look less pretty, but it may help when exporting html-tables to office tools.

Value

Invisibly returns

the web page style sheet (page.style),

the web page content (page.content),

the complete html-output (page.complete) and

the html-table with inline-css for use with knitr (knitr)

for further use.

Note

Computation of p-values (if necessary) is based on normal-distribution assumption, treating the t-statistics as Wald z-statistics.

The variance components of the random effects (see show.re.var) are denoted like:

between-group-variance: tau-zero-zero

random-slope-intercept-correlation: rho-zero-one

Standard errors for generalized linear (mixed) models are not the regular standard errors on the untransformed scale, as shown in the summary()-method. Rather, sjt.glmer() uses adjustments according to the delta method for approximating standard errors of transformed regression parameters (see se).